Move alerts to own handler
