+++ app/src/handlers/index.ts
@@ -2,7 +2,10 @@ import { ContainerModule } from 'inversify';
 import PlayerHandlerInterface from '../interfaces/PlayerHandlerInterface';
 import PlayerHandler from './PlayerHandler';
 import {TYPES} from '../utils/types';
+import AlertHandlerInterface from '../interfaces/AlertHandlerInterface';
+import AlertHandler from './AlertHandler';
 
 export default new ContainerModule(bind => {
     bind<PlayerHandlerInterface>(TYPES.PlayerHandlerInterface).to(PlayerHandler).inSingletonScope();
+    bind<AlertHandlerInterface>(TYPES.AlertHandlerInterface).to(AlertHandler).inSingletonScope();
 });
