+++ app/src/interfaces/AlertHandlerInterface.ts
@@ -1,0 +1,5 @@
+import MetagameEventEvent from '../handlers/census/events/MetagameEventEvent';
+
+export default interface AlertHandlerInterface {
+    handleMetagameEvent(mgeEvent: MetagameEventEvent): boolean;
+}
