+++ app/src/utils/types.ts
@@ -1,6 +1,8 @@
 const TYPES = {
     // eslint-disable-next-line @typescript-eslint/naming-convention
     PlayerHandlerInterface: Symbol.for('PlayerHandlerInterface'),
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    AlertHandlerInterface: Symbol.for('AlertHandlerInterface'),
 };
 
 export { TYPES };
